[connectors] - refactor: move whitelist channel patterns function to index
 - Extracted `getWhiteListedChannelPatterns` function from `slack/bot.ts` to `slack/index.ts`
 - Ensured fetching of whiteListedChannelPatterns from Slack configuration remains functional after the move
